Move wrapper functions into Raylib.Utils
